Alg works!!
I was having issues where the errors would suddenly increase after the 10-14th generation. We figured out it was due to the lack of actual grades which are a "D"(the alg was evolving to not use that terminal). instead of changing how we assign the grade boundaries me and Nic upped the population size from 300 to 10000.